return map values at double precision

+++ python/wssa_utils.py
@@ -65,7 +65,7 @@ def tile_par_struc(large=True, release='1.0'):
     # conversion from factor from W3 DN to MJy/sr
     calfac = 0.0163402
     # data type for intensity and bit-mask outputs, respectively
-    dtype  = ['float32', 'int32']
+    dtype  = ['float64', 'int32']
 
     par = {'sidelen'   : sidelen,
            'pix'       : pix,
